fix: rename components for front-layout
