+++ src/views/FrontLayoutView.vue
@@ -2,16 +2,16 @@
 
 <template>
   <div id="front">
-    <FontHeader />
+    <FrontHeader />
     <main class="container">
       <RouterView />
     </main>
-    <FontFooter />
+    <FrontFooter />
   </div>
 </template>
 
 <script setup>
 import { RouterView } from 'vue-router'
-import FontHeader from '@/components/front/base/FontHeader.vue'
-import FontFooter from '@/components/front/base/FontFooter.vue'
+import FrontHeader from '@/components/front/base/FrontHeader.vue'
+import FrontFooter from '@/components/front/base/FrontFooter.vue'
 </script>
